Add files via upload
# Project Code

This repository contains the complete code for the Smart Street Light System. Developed during my second year of undergraduate studies, the code is designed to automatically control the brightness of street lights based on ambient sunlight intensity.

## Overview

The project utilizes an LDR sensor, a NodeMCU ESP8266 microcontroller, and an LED street light. The system adjusts the brightness of the street light in real-time according to the data received from the LDR sensor. Additionally, it integrates with the ThingSpeak platform for remote monitoring and visualization.

## Code Highlights

- **Sensor Data Acquisition**: The LDR sensor measures ambient light intensity and sends this data to the microcontroller.
- **Brightness Control Algorithm**: The NodeMCU ESP8266 processes the sensor data and adjusts the LED street light brightness accordingly.
- **ThingSpeak Integration**: The microcontroller sends data to ThingSpeak for real-time monitoring and visualization.
- **Energy Efficiency**: The code ensures optimal brightness levels, reducing energy consumption during periods of sufficient natural light.

## How to Use

1. **Clone the Repository**:
   ```sh
   git clone https://github.com/Satwikr15/SmartStreetLightSystem.git